requirements.txt
aiogram==3.0.0
openai==1.0.0
python-docx==0.8.11
pydantic==1.9.0
python-dotenv==0.20.0
